Read cells from an ODS sheet

Import cells from an OpenDocument Spreadsheet (ODS) file. In the resulting dataaset each cell is its own row with columns providing information about the cell's postion, value types and containing formulas.

Usage

read_ods_cells(path, sheet)

Arguments

path

The ODS file

sheet

The sheet within the ODS file

Value

A tibble with the following information:

row: the row number of the cell

col: the column number of the cell

cell_type: the type of cell, either "cell", "empty" or "merged"

value_type: the ODS value type (see details)

cell_formula: the ODS formula specification

cell_content: the value of the cell as represented to the user of a spreadsheet application (i.e. with text formatting applied)

base_value: the underlying value of a cell (see details)

currency symbol: the currency symbol for cells with a value_type of currency

Details

Cells are assigned a cell_type of "cell" when they have content, "empty" if they have no content, or "merged" if they are part of a set of merged cells.

The value_type of a cell is based on the "value-type" attribute defined by the OpenDocument Format specification. There are 7 value types:

boolean: equivalent to R's base::logical() data type

currency: a numeric value, optionally with a currency symbol

date: a date or date-time stored in ISO format

float: a numeric value, the ODS format does not distinguish between different numeric types

percentage: a numeric value, formatted for display as a percentage

string: equivalent to R's base::character()

time: a duration, stored in ISO format

For currency value types optionally a currency symbol (e.g. GBP, EUR, USD) can be set, this is provided separate from the base_value as currency_symbol.

read_ods_cells provides two representations of a cell's value. The cell_content column provides the cell's value as presented to the user of a traditional spreadsheet application (such as Microsoft Excel, Google Sheets or LibreOffice Calc), i.e. after applying cell formatting rules. The base_value column provides the cell's underlying value, i.e. without formatting.