Read cells from an ODS sheet

Import cells from an OpenDocument Spreadsheet (ODS) file. In the resulting dataset each cell is its own row with columns providing information about the cell's position, value types and containing formulas.

Usage

read_ods_cells(
  path,
  sheet = 1,
  quick = FALSE,
  whitespace = FALSE,
  quiet = FALSE
)

Arguments

path

The ODS file

sheet

The sheet name or index number, set to NA for all sheets

quick

Whether to use the quick reading process

whitespace

Whether to process multiple whitespaces (ignore for now)

quiet

Whether to silence console messages (recommended for bulk processing)

Value

A tibble (data.frame) of cells from the ODS sheet(s).

Details

The goal of read_ods_cells() is to extract the constituent value(s) and other information about cells stored in an ODS file and to present that in a "tidy" format that allows further programmatic manipulation. It is modelled after the functionality of tidyxl::tidy_xlsx() which performs a similar role for Microsoft Excel spreadsheets.

There are between two to four presentations of a cell's value in the resulting tibble:

the base_value, a character vector providing the "most raw" version of a cell's value (see below regarding variation when setting quick=TRUE);

the cell_content, a character vector providing the version of a cell's value as seen by the user of a spreadsheet application;

for float, currency percentage value types, a numeric_value with the raw value of the cell as a base R numeric vector, currency value types also have a currency_symbol providing the 3-character ISO currency symbol;

for date and time value types, a character vector with the raw value conforming to the relevant ISO standard.

Processing the ODS XML is a memory intensive process, you can achieve significant speed enhancements by setting the quick argument to FALSE. This process wille extract only a minimum of information about the cells, namely: sheet, row, col, value_type and a base_value. The base_value when using the quick argument will combine the raw value stored for float and percentage value types with the cell_content (i.e formatted character string) for all other value types.

#' @details # Cell information The resulting data.frame has 25 columns:

sheet: the sheet the cell is from

row: the row number of the cell

col: the column number of the cell

cell_type: the type of cell, either cell, empty, merge-hidden or merge-lead

is_empty: a logical vector indicating if the cell is empty

value_type: the value type, either boolean, currency, date, float, percentage, string or time

cell_content: the content as presented to users of a spreadsheet application

base_value: the "base" value of the cell

numeric_value: base values for numeric value types as a numeric vector

currency_symbol: the currency symbol associated with currency value types

boolean_value: base values for boolean value types as a logical vector

date_value: base values for date value types (incl date-time) as a character vector

time_value: base value for time value types as a character vector

has_formula: a logical vector indicating if the cell content are calculated by a formula

formula: the formula used to calculate the cell content

has_error: a logical vector indicating if the cell formula results in an error

error_type: a numeric vector indicating the type of error

has_annotation: a logival vector indicating if the cell has a comment or annotation

annotation: the text of any comment or annotation

is_merged: a logical indicating if the cell is part of a merge group

merge_colspan: for merge leader cells, the number of columns spanned

merge_rowspan: for merge leader cells, the number of rows spanned

merge_shape: for merge leader cells, the shape of the merge group, either vetical, horizontal, or rectangle

cell_style: the style code of the cell

row_style: the style code of the row

Cell types and merge groups

Cells are assigned one of four cell_type values:

a cell is empty if it has no content or no value type,

a cell is a merge-lead if it is the top-left cell of a merge group, other cells in the merge group are assigned merge-hidden

everything else is simply a cell

Cells in a merge group will also have is_merged set to TRUE. The top-left cell in a merge group has information about the number of columns (merge_colspan) and rows (merge_rowspan) in the merge group as well as an indication of shape (vertical, horizontal or rectangle). Merge groups are not uniquely identified.

The ODS specification allows for cells hidden by a merge to have content, annecdotally this appears to be supported by ODS files created by LibreOffice but not by ODS files created by Microsoft Excel.

Value types

The value_type of a cell is based on the "value-type" attribute defined by the OpenDocument Format specification. There are 7 value types:

boolean: equivalent to R's base::logical() data type

currency: a numeric value, optionally with a currency symbol

date: a date or date-time stored in ISO format

float: a numeric value, the ODS format does not distinguish between different numeric types

percentage: a numeric value, formatted for display as a percentage

string: equivalent to R's base::character()

time: a duration, stored in ISO format

For currency value types optionally a currency symbol (e.g. GBP, EUR, USD) can be set, this is provided separately from the base_value as currency_symbol, the formatting of currency symbols is handled by style information, the value as visible to a spreadsheet user is provided in the cell_content column.

Error types

The values of error_type are equivalent to those produced by Microsoft Excel and LibreOffice.

Examples

example <- system.file("extdata", "basic_example.ods", package = "tidyods")
example_cells <- read_ods_cells(example, 1)
#> ℹ Extracting XML
#> ✔ Extracting XML [26ms]
#> 
#> ℹ Getting sheet
#> ✔ Getting sheet [17ms]
#> 
#> ℹ Extracting cell and row info
#> ✔ Extracting cell and row info [144ms]
#> 
dplyr::glimpse(example_cells)
#> Rows: 28
#> Columns: 25
#> $ sheet           <chr> "penguins", "penguins", "penguins", "penguins", "pengu…
#> $ row             <dbl> 1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3, 4, 4, 4, 4, 5, 5, …
#> $ col             <dbl> 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4, 1, 2, …
#> $ cell_type       <chr> "cell", "cell", "cell", "cell", "cell", "cell", "cell"…
#> $ is_empty        <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ value_type      <chr> "string", "string", "string", "string", "string", "boo…
#> $ cell_content    <chr> "species", "female", "bill_length_mm", "body_mass_g", …
#> $ base_value      <chr> "species", "female", "bill_length_mm", "body_mass_g", …
#> $ numeric_value   <dbl> NA, NA, NA, NA, NA, NA, 40.39041, 4043.49315, NA, NA, …
#> $ currency_symbol <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ boolean_value   <lgl> NA, NA, NA, NA, NA, FALSE, NA, NA, NA, TRUE, NA, NA, N…
#> $ date_value      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ time_value      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ has_formula     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ formula         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ has_error       <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ error_type      <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ has_annotation  <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ annotation      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ is_merged       <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE…
#> $ merge_colspan   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ merge_rowspan   <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ merge_shape     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA…
#> $ cell_style      <chr> NA, "Default", "Default", "Default", NA, NA, NA, NA, N…
#> $ row_style       <chr> "ro1", "ro1", "ro1", "ro1", "ro1", "ro1", "ro1", "ro1"…